feat: handle printing
* add termion dependency

* add run function to Explorer with loop inside

* add refresh_screen function to Explorer

* add handle_keypress function to Explorer (only handles quit atm)

* add Terminal struct

+++ src/cursor.rs
@@ -1,7 +1,7 @@
 use crate::directory::Directory;
 
 pub struct Cursor {
-    pub position: i32,
+    pub position: usize,
     pub max: i32,
 }
